Avoid sending MQTT messages before CONNACK has been received
This fixes the problem with AWS broker which doesn't accept SUBSCRIBE packets before the CONNACK packet has been received, instead returning UNSPECIFIED_ERROR in SUBACK. This is not standard compliant which allows the client to send control packets right after CONNECT, before CONNACK has been received. But with the new behaviour we increase compatibility.
This problem manifests itself mainly with the Android Emulator.